Fix a bug in JPlaceholder's placeholder keypair extractor that incorrectly assummed non-placeholder suffix chars to be part of the placeholder
Example: A placeholder like "${regex ^image([-_0-9]{0,})$}" would be seen as "regex = ^image([-_0-9]{0,", which is incorrect, as the first closing curly brace is part of the regex, not the closing of the placeholder. This new method instead correctly pulls out the placeholder data so it looks like "regex = image([-_0-9]{0,})$". This new method is backwards compatible with the old one and is a drop-in replacement, hence why no deprecation occurred.